Fixed a bug where the prop firm connection was not being checked.
mt5 keeps the same connection open for all prop firms. This was causing issues
where the wrong account was being used.
This change checks if the account info is for the same user as the one in the
credentials. If not, it shuts down the connection and opens a new one.